requirements.txt
asgiref==3.7.0
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
drf-nested-routers==0.93.4
fonttools==4.39.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.0
asgiref==3.7.0
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
drf-nested-routers==0.93.4
fonttools==4.39.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.0
mysql==0.0.3
mysqlclient==2.1.1
tzdata==2023.3
django-cors-headers==4.1.0
